Remove old page and components
