--- src/styles/components/tabs.scss
@@ -1,84 +1,0 @@
-$tabs-border-bottom-color: $color-light-gray;
-$tabs-border-bottom-width: 0.125rem;
-$tabs-font-size: 1rem;
-
-.tabs {
-  display: flex;
-  flex-direction: column;
-
-  justify-content: space-between;
-  overflow: hidden;
-  overflow-x: auto;
-  font-family: $family-inter;
-  [role='tablist'] {
-    display: flex;
-    border-bottom: $tabs-border-bottom-width solid $tabs-border-bottom-color;
-    font-size: 1rem;
-  }
-  .tabs-panel {
-    display: none;
-    padding: 2rem 0 0;
-    &.is-active {
-      display: block;
-    }
-  }
-
-  .tab {
-    appearance: none;
-    border-image-source: unset;
-    border-image-outset: unset;
-    border-image-repeat: unset;
-    font-weight: 700;
-    margin: 0 1rem -0.125rem 1rem;
-    padding: 0.5rem 0 1rem;
-    position: relative;
-    border-width: 0;
-    background: transparent;
-    border-bottom: 0.25rem solid transparent;
-    color: #333;
-    cursor: pointer;
-    user-select: none;
-    text-decoration: none;
-
-    &.is-active,
-    &:hover {
-      border-bottom-color: #333;
-      text-decoration: none;
-    }
-    &:focus {
-      outline: none;
-    }
-    &:focus-visible {
-      outline: none;
-      box-shadow: 0 0 0 0.125rem #d8d8d8;
-    }
-  }
-}
-.boxed {
-  [role='tablist'] {
-    display: flex;
-    align-self: flex-start;
-    border-bottom: none;
-  }
-  .tabs-panel {
-    padding: 1rem;
-    background-color: white;
-    border: 1px solid $color-light-gray;
-  }
-  .tab {
-    font-size: 1rem;
-    margin: 0 0 -0.125rem;
-    border: 1px solid $color-light-gray;
-    padding: 0.5rem 1rem;
-
-    &.is-active,
-    &:hover {
-      background-color: white;
-      border-bottom: 0;
-      padding-bottom: 0.625rem;
-    }
-    &:focus-visible {
-      box-shadow: 0 0 0 0.125rem #d8d8d8;
-    }
-  }
-}
