+++ src/styles/vocabulary.scss
@@ -21,6 +21,5 @@
 @import 'components/navbar';
 @import 'components/footer';
 @import 'components/card';
-@import 'components/tabs';
 @import 'components/table';
 @import 'components/content';
